Support for correlating episodes, i.e. support the case of not generating one transaction per episode.
This will result in many "useless patterns", in that they contain "noisy" information about correlated episode durations. The upside, however, is that we *can* correlate them. Or we can filter this away by simply only accepting *one* episode per pattern — after all, if a pattern is deemed frequent enough *with* a correlated episode, it must also be without, and hence we don't ignore anything.

This may break the RuleMiner though, since it has some optimization on the assumption that there is one transaction per episode.